Improve briefing text handling a bit
This now creates elements and text elements directly instead of relying
on setting an RML string to the parent. This should be a bit easier to
extend in the future when color handling is added and also more robust
since it does not require escaping the text since it is added as a text
node which does not need escaping to work.