Fix two more barracks issues
Found two more small issues with Barracks. One is that SCPUI allows the player to press escape and exit to mainhall when all players are deleted, though I have fixed this with just an early return.

The other one is that SCPUI does not need to specify a default multiplayer logo since that will throw errors for mods in which it does not exist. Removing the default in the RML has no effect sine the lua file builds the lists dynamically anyway.

Tested and both work as expected.