Dialog order consistency
Make all dialogs consistently show `cancel/no` as top choice then `yes/confirm` as bottom choice. Most SCPUI already have this pattern but I found 4 that do not. (a bit backwards since to have the order in the dialog be No, Yes the order in the code needs to be Yes, No.